- changed how timer is implemented

+++ MyGame/src/Object/RigidBody.h
@@ -4,12 +4,12 @@ class RigidBody
 {
 public:
 	RigidBody();
-	inline int GetPositionX() { return xPos; }
-	inline int GetPositionY() { return yPos; }
-	inline void MovePosX(int x) { xPos += x; }
-	inline void MovePosY(int y) { yPos += y; }
+	inline int GetPositionX() { return (int)xPos; }
+	inline int GetPositionY() { return (int)yPos; }
+	inline void MovePosX(float x) { xPos += x; }
+	inline void MovePosY(float y) { yPos += y; }
 private:
-	int xPos,yPos;
+	float xPos,yPos;
 };
 
 #endif
